+++ CHANGELOG.md
@@ -19,6 +19,23 @@ All notable changes to this project will be documented in this file.
 
 - more performant query for `blockchain.master_seq_no_range.end`
 
+### Removed
+
+Queries:
+- `blockchain.workchain_blocks`. Use `blockchain{ blocks }` instead.
+- `blockchain.workchain_transactions`. Use `blockchain{ transactions } ` instead.
+- `blockchain.account_transactions`. Use `blockchain{ account{ transactions } }` instead.
+- `explainQueryAccounts`
+- `explainQueryTransactions`
+- `explainQueryMessages`
+- `explainQueryBlocks`
+- `explainQueryBlockSignatures`
+- `explainQueryZerostates`
+- `getAccountsCount`
+- `getTransactionsCount`
+- `getAccountsTotalBalance`
+- `QueryExplanation` and `SlowReason` types
+
 ## [0.51.2] - 2022-05-10
 
 ### New